Support single- and multiple-target VTS's.
Each target gets wrapped in a VersionedTargetSet, but we rely on
the assumption that it is alone in its set until later, when we
group the single versioned targets into multi-target sets for
compilation.

This makes this distinction more explicit and allows for more
complicated logic elsewhere about which targets to group
together.